feat(frontend): add reading progress and hero to post page
